TargetingCalculator determines how to shoot from a given point on the field.
The purpose of this commit is to provide an example of what readable Java
code looks like and a way to test it. It is primarily intended to be
instructional.

Note that the actual interpolation math is not implemented. While critical,
the details are needed to illustrate what code and tests can look like.
How to add the calculations into the calculator should be obvious (not
necessarily what those calculations look like, which is an excersize for
the reader).

See the TargetingCalulator class for more details about how it works and
how to use it.  Note that the Point and TargetingParams classes are also part
of the public interface required for use.

Ideally the Point class should be some pre-existing class we are using for
"x, y" field locations. However I could not find what this would be so
introduced one. If there is no such class then Point belongs somewhere
more appropriate, such as in the lib.util package.

The TargetingParams class contains the parameters needed to configure
the robot to score from a given point on the field. Pre-calibrated parameters
for given points on the field are provided to the TargetingCalculator. It
will use these to determine whhat the parameters should be for any given
point on the field within shooting range.

The tests provide additional documentation and usage information. Not
necessarily in comments, but through the test cases and hopefully readable
code. Sometimes tests may cheat and access private information within the
components so that it can easily and clearly setup certain interesting
scenarios to test which would either be complicated or very subtle to
understand what's being tested. However on the most part the tests follow
normal usage patterns.